% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testcppautocomplete-package.R
\docType{package}
\name{testcppautocomplete-package}
\alias{testcppautocomplete}
\alias{testcppautocomplete-package}
\title{testcppautocomplete: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/testcppautocomplete}
  \item Report bugs at \url{https://github.com/DavisVaughan/testcppautocomplete/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
